dbus: automatically dismiss gnome-online-accounts's login keyring unlock request
We now unlock the keyring whenever the user inputs their PIN for the first time,
which is enough to keep GOA content. Unfortunately the popup asking the user for
their password does not automatically go away once the login keyring is unlocked
so this ugly hack manually takes care of it. Not my finest work, but...